optional tmp clear

+++ synthesis.sh
@@ -6,9 +6,19 @@ cesta_vstup=$1
 # Jazyk modelu pro syntézu je druhým parametrem
 jazyk=$2
 
+vycistit_tmp=$3
+
+if [ $# -gt 0 ] && [ "$cesta_vstup" = "help" ]; then
+  echo "Usage: ./synthesis.sh <input_file> <language> [rm]"
+  echo "  <input_file>  - path to the text file with the book"
+  echo "  <language>    - language of the book (en/cs)"
+  echo "  [rm]          - optional parameter to remove temporary files"
+  exit 0
+fi
+
 # Ověřím správnost zadaných parametrů
 if [ $# -lt 2 ]; then
-    echo "Usage: ./synthesis.sh <language> <input_file>" >&2
+    echo "Usage: ./synthesis.sh <input_file> <language> [rm]" >&2
     exit 1
 elif [ ! -f "${cesta_vstup}" ]; then
     echo "Input file ${cesta_vstup} does not exist." >&2
@@ -51,14 +61,8 @@ else
   sleep 5 # čekám 5 sekund, než se server zapne
 
   # Připrav příkaz pro syntézu v angličtině
-  # https://community.openconversational.ai/t/encoding-issue-with-mimic3-server-latin-1-vs-utf-8/13980
   cmd='echo "${radky[$i]}" | iconv -f UTF-8 -t ISO-8859-1//TRANSLIT 2>/dev/null | mimic3 --remote > "$TMPDIR/audio${i}.wav"'
 fi
-
-# optional feature: paralelizace ??
-#cmd=$cmd' && echo "file $TMPDIR/audio${i}.wav" >> $TMPDIR/concat.txt'
-# pokud paralelizace, tak zde
-# cmd=$cmd' &'
 
 # Procházím pole položku po položce (procházím text řádku po řádce)
 for (( i=0; i<${#radky[@]}; i++ )); do
@@ -78,4 +82,6 @@ fi
   #ffmpeg -f concat -safe 0 -i $TMPDIR/concat.txt -c copy $cesta_vystup
   # S relativními cestami uvnitř concat.txt nemusím používat -safe 0
 ffmpeg -f concat -i $TMPDIR/concat.txt -c copy $cesta_vystup
-rm -rf $TMPDIR/ 2>/dev/null
+if [ $vycistit_tmp ]; then
+  rm -rf $TMPDIR/ 2>/dev/null
+fi
